requirements.txt
openai>=1.3.0
pyaudio>=0.2.13
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pillow>=10.0.0
mss>=9.0.1
python-dotenv>=1.0.0
numpy>=1.24.0
